requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
GitPython